LayoutTests/http/wpt/navigation-api/transition-promises-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error

PASS NavigationTransition finished promise is fulfilled on success
PASS NavigationTransition finished promise is rejected on error
